apps/text-to-audio/requirements.txt
streamlit
azure-cognitiveservices-speech
langchain
langchain-community
pypdf
requests
beautifulsoup4
python-dotenv
